Finalizando a primeira aula

+++ src/main/java/br/com/alura/agencias/domain/Agencia.java
@@ -14,7 +14,7 @@ public class Agencia {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	private String nome;
 
 	@Column(name = "razao_social")
@@ -29,7 +29,7 @@ public class Agencia {
 
 	}
 
-	public Agencia(Integer id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
+	public Agencia(Long id, String nome, String razaoSocial, String cnpj, Endereco endereco) {
 		this.id = id;
 		this.nome = nome;
 		this.razaoSocial = razaoSocial;
@@ -37,7 +37,7 @@ public class Agencia {
 		this.endereco = endereco;
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
